+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 current dev branch
 
+### Added
+
+- utils
+    - colors
+    
 ## [0.4.1]  - 17-01-2020
 
 ### Fixed